re-introduce task counts
