+++ package.json
@@ -30,7 +30,7 @@
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.2.2",
     "react-router-redux": "next",
-    "react-scripts": "1.0.17",
+    "react-scripts": "1.1.1",
     "react-scroll-to-component": "^1.0.2",
     "react-test-renderer": "^16.2.0",
     "react-transition-group": "^2.2.1",
